vector_poc_tests/config.py
"""
Central configuration — reads from .env and exposes typed settings.

Updated to use VectorDBBench real-world dataset (Performance1536D500K)
instead of synthetic random vectors.
"""
from __future__ import annotations
import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv(Path(__file__).parent / ".env")


def _int(key: str, default: int) -> int:
    return int(os.getenv(key, str(default)))


def _float(key: str, default: float) -> float:
    return float(os.getenv(key, str(default)))


def _list_int(key: str, default: str) -> list[int]:
    return [int(x.strip()) for x in os.getenv(key, default).split(",")]


# -- Connection ---------------------------------------------------------------
def get_dsn() -> str:
    """Build a PostgreSQL DSN from env vars."""
    dsn = os.getenv("CRDB_DSN")
    if dsn:
        return dsn
    host = os.getenv("CRDB_HOST", "localhost")
    port = os.getenv("CRDB_PORT", "26257")
    user = os.getenv("CRDB_USER", "root")
    pw = os.getenv("CRDB_PASSWORD", "")
    db = os.getenv("CRDB_DATABASE", "vectortest")
    ssl = os.getenv("CRDB_SSLMODE", "verify-full")
    cert = os.getenv("CRDB_SSLROOTCERT", "")
    dsn = f"postgresql://{user}:{pw}@{host}:{port}/{db}?sslmode={ssl}"
    if cert:
        dsn += f"&sslrootcert={cert}"
    return dsn


# -- Vector / Dataset ---------------------------------------------------------
VECTOR_DIM: int = _int("VECTOR_DIM", 1536)

# VectorDBBench dataset: Performance1536D500K (OpenAI embeddings from C4 corpus)
# This replaces synthetic random vectors with real-world embeddings that have
# natural cluster structure, giving production-representative results.
DATASET_NAME: str = os.getenv("DATASET_NAME", "Performance1536D500K")
DATASET_DIR: str = os.getenv("VECTORDBBENCH_DATA_DIR", "")  # override dataset location

# Subset sizes carved from the 500K train set for scaling tests
DATASET_SIZES: list[int] = _list_int("DATASET_SIZES", "50000,100000,250000,500000")
SMALL_DATASETS: list[int] = [s for s in DATASET_SIZES if s <= 250_000]
LARGE_DATASET: int = max(DATASET_SIZES)

# -- C-SPANN Index Tuning -----------------------------------------------------
# beam_size controls recall vs. latency trade-off.
#   - Default 32: ~14% recall on RANDOM vectors, but much higher on real data
#     with natural cluster structure (which is what production looks like)
#   - 128: ~90% recall on real-world Cohere 1M benchmark
#   - 2048: ~94.7% recall on random vectors (unrealistic for production)
#
# Using default beam_size=32 with the real-world VectorDBBench dataset
# demonstrates that production-like data doesn't need aggressive tuning.
# The cluster structure in real embeddings means the index efficiently
# narrows the search space even at low beam sizes.
VECTOR_SEARCH_BEAM_SIZE: int = _int("VECTOR_SEARCH_BEAM_SIZE", 32)
BUILD_BEAM_SIZE: int = _int("BUILD_BEAM_SIZE", 8)
MIN_PARTITION_SIZE: int = _int("MIN_PARTITION_SIZE", 16)
MAX_PARTITION_SIZE: int = _int("MAX_PARTITION_SIZE", 128)

# -- Test Params ---------------------------------------------------------------
QUERY_COUNT: int = _int("QUERY_COUNT", 1000)
TOP_K: int = _int("TOP_K", 10)              # K=10 per Soma's recommendation
K_VALUES: list[int] = [1, 5, 10]
ITERATIONS: int = _int("ITERATIONS", 3)
BATCH_SIZE: int = _int("BATCH_SIZE", 500)
CONCURRENCY_LEVELS: list[int] = _list_int("CONCURRENCY_LEVELS", "1,5,10")
FRESHNESS_TRIALS: int = _int("FRESHNESS_TRIALS", 100)

# -- Thresholds (updated for real-world dataset) ------------------------------
# With real-world embeddings (clustered data), C-SPANN performs significantly
# better than with random vectors. Thresholds adjusted accordingly.
QUERY_P95_SMALL_MS: float = _float("QUERY_P95_SMALL_MS", 100.0)   # p95 < 100 ms for 50K-250K
QUERY_P95_LARGE_MS: float = _float("QUERY_P95_LARGE_MS", 200.0)   # p95 < 200 ms for 500K
QUERY_P99_SMALL_MS: float = _float("QUERY_P99_SMALL_MS", 150.0)   # p99 < 150 ms for 50K-250K
QUERY_P99_LARGE_MS: float = _float("QUERY_P99_LARGE_MS", 300.0)   # p99 < 300 ms for 500K
INSERT_AVG_MS: float = _float("INSERT_AVG_MS", 50.0)              # avg insert/update < 50 ms
BATCH_THROUGHPUT_MIN: int = _int("BATCH_THROUGHPUT_MIN", 5000)     # >= 5K embeddings/min
FRESHNESS_P95_SEC: float = _float("FRESHNESS_P95_SEC", 5.0)       # searchable < 5 sec
RECALL_AT_K: float = _float("RECALL_AT_K", 0.90)                  # recall@10 >= 90%
STORAGE_MAX_GB: float = _float("STORAGE_MAX_GB", 5.0)             # < 5 GB for 500K vectors
STORAGE_EXPECTED_MIN_GB: float = 3.0
STORAGE_EXPECTED_MAX_GB: float = 4.0

# -- Paths --------------------------------------------------------------------
PROJECT_DIR = Path(__file__).parent
REPORTS_DIR = PROJECT_DIR / "reports"
DATA_DIR = PROJECT_DIR / "data"
REPORTS_DIR.mkdir(exist_ok=True)
DATA_DIR.mkdir(exist_ok=True)
