vector_poc_tests/requirements.txt
psycopg[binary]>=3.1
psycopg_pool>=3.1
numpy>=1.24
python-dotenv>=1.0
rich>=13.0
jinja2>=3.1
click>=8.1
tabulate>=0.9
